Fix errors after moving from a solochain
