+++ pallets/parachain-utils/Cargo.toml
@@ -17,8 +17,6 @@ std = [
     'scale-info/std',
     'frame-support/std',
     'frame-system/std',
-    'sp-runtime/std',
-    'sp-std/std',
 ]
 
 [dependencies]
@@ -30,9 +28,3 @@ jsonrpc-core = '18.0.0'
 # Substrate dependencies
 frame-support = { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.12', default-features = false }
 frame-system = { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.12', default-features = false }
-sp-runtime = { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.12', default-features = false }
-sp-std = { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.12', default-features = false }
-
-[dev-dependencies]
-sp-core = { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.12', default-features = false }
-sp-io = { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.12', default-features = false }
