--- pallets/parachain-utils/src/mock.rs
@@ -1,80 +1,0 @@
-use super::*;
-
-use sp_core::H256;
-use sp_io::TestExternalities;
-
-use sp_runtime::{
-    traits::{BlakeTwo256, IdentityLookup}, testing::Header,
-};
-use frame_support::{parameter_types, traits::Everything};
-
-use crate as pallet_parachain_utils;
-use crate::{DEFAULT_MIN_HANDLE_LEN, DEFAULT_MAX_HANDLE_LEN};
-
-type UncheckedExtrinsic = frame_system::mocking::MockUncheckedExtrinsic<Test>;
-type Block = frame_system::mocking::MockBlock<Test>;
-
-frame_support::construct_runtime!(
-    pub enum Test where
-        Block = Block,
-        NodeBlock = Block,
-        UncheckedExtrinsic = UncheckedExtrinsic,
-    {
-        System: frame_system::{Pallet, Call, Config, Storage, Event<T>},
-        ParachainUtils: pallet_parachain_utils::{Pallet},
-    }
-);
-
-parameter_types! {
-    pub const BlockHashCount: u64 = 250;
-    pub BlockWeights: frame_system::limits::BlockWeights =
-        frame_system::limits::BlockWeights::simple_max(1024);
-}
-impl frame_system::Config for Test {
-    type BaseCallFilter = Everything;
-    type BlockWeights = ();
-    type BlockLength = ();
-    type Origin = Origin;
-    type Call = Call;
-    type Index = u64;
-    type BlockNumber = u64;
-    type Hash = H256;
-    type Hashing = BlakeTwo256;
-    type AccountId = AccountId;
-    type Lookup = IdentityLookup<Self::AccountId>;
-    type Header = Header;
-    type Event = Event;
-    type BlockHashCount = BlockHashCount;
-    type DbWeight = ();
-    type Version = ();
-    type PalletInfo = PalletInfo;
-    type AccountData = ();
-    type OnNewAccount = ();
-    type OnKilledAccount = ();
-    type SystemWeightInfo = ();
-    type SS58Prefix = ();
-    type OnSetCode = ();
-}
-
-parameter_types! {
-    pub const MinHandleLen: u32 = DEFAULT_MIN_HANDLE_LEN;
-    pub const MaxHandleLen: u32 = DEFAULT_MAX_HANDLE_LEN;
-  }
-impl Config for Test {}
-
-pub type AccountId = u64;
-
-pub struct ExtBuilder;
-
-impl ExtBuilder {
-    pub fn build() -> TestExternalities {
-        let storage = frame_system::GenesisConfig::default()
-            .build_storage::<Test>()
-            .unwrap();
-
-        let mut ext = TestExternalities::from(storage);
-        ext.execute_with(|| System::set_block_number(1));
-
-        ext
-    }
-}
